CC/app.py
# Import necessary libraries
from flask import Flask, request, jsonify
from google.cloud import storage
import os

app = Flask(__name__)


# Define a route for processing image-scanning requests
@app.route('/scan-image', methods=['POST'])
def scan_image():
    try:
        # Check if an image file is present in the request
        if 'image' not in request.files:
            return jsonify({'error': 'No image file provided'}), 400

        image_file = request.files['image']

        # Upload the image to Cloud Storage
        image_url = upload_to_cloud_storage(image_file)

        # Implement your machine learning logic here for image processing

        # Return the result or any relevant information
        return jsonify({'image_url': image_url, 'result': 'Image scanned successfully'}), 200

    except Exception as e:
        return jsonify({'error': str(e)}), 500

def upload_to_cloud_storage(file):
    # Configure Google Cloud Storage client
    storage_client = storage.Client()
    bucket_name = 'your-cloud-storage-bucket'

    # Upload the file to Cloud Storage
    bucket = storage_client.bucket(bucket_name)
    blob = bucket.blob(file.filename)
    blob.upload_from_file(file)

    # Generate a public URL for the uploaded file
    image_url = f'https://storage.googleapis.com/{bucket_name}/{file.filename}'

    return image_url

if __name__ == '__main__':
    # Run the Flask application
    app.run(host='0.0.0.0', port=int(os.environ.get('PORT', 8080)))
