CC/requirements.txt
Flask==2.0.1
google-cloud-storage==1.44.0
tensorflow==2.8.0
# Add other libraries as needed
